requirements.txt
asgiref==3.3.0
bson==0.5.8
dataclasses==0.6
dj-database-url==0.5.0
Django==2.2.17
django-fs-trumbowyg==0.1.4
django-rest-framework==0.1.0
django-tinymce==3.1.0
django-utils==0.0.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.5.0
djongo==1.3.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
pymongo==3.11.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.2.4
